+++ app/src/main/java/elite/intel/ai/brain/handlers/commands/LightsOnOffHandler.java
@@ -12,41 +12,13 @@ public class LightsOnOffHandler extends CommandOperator implements CommandHandle
 
     @Override public void handle(String action, JsonObject params, String responseText) {
         Status status = Status.getInstance();
-        boolean on = params.get("state").getAsBoolean();
 
         if (status.isInSrv()) {
-            if (on) {
-                if (status.isLightsOn()) {
-                    return;
-                } else {
-                    toggleLights(Bindings.GameCommand.BINDING_BUGGY_LIGHTS_TOGGLE.getGameBinding());
-                }
-            } else {
-                if (!status.isLightsOn()) {
-                    return;
-                } else if (status.isSrvHighBeam()) {
-                    toggleLights(Bindings.GameCommand.BINDING_BUGGY_LIGHTS_TOGGLE.getGameBinding());
-                } else {
-                    toggleLights(Bindings.GameCommand.BINDING_BUGGY_LIGHTS_TOGGLE.getGameBinding());
-                    toggleLights(Bindings.GameCommand.BINDING_BUGGY_LIGHTS_TOGGLE.getGameBinding());
-                }
-            }
+            toggleLights(Bindings.GameCommand.BINDING_BUGGY_LIGHTS_TOGGLE.getGameBinding());
         }
 
         if (status.isInMainShip()) {
-            if (on) {
-                if (status.isLightsOn()) {
-                    return;
-                } else {
-                    toggleLights(Bindings.GameCommand.BINDING_SHIP_LIGHTS_TOGGLE.getGameBinding());
-                }
-            } else {
-                if (!status.isLightsOn()) {
-                    return;
-                } else {
-                    toggleLights(Bindings.GameCommand.BINDING_SHIP_LIGHTS_TOGGLE.getGameBinding());
-                }
-            }
+            toggleLights(Bindings.GameCommand.BINDING_SHIP_LIGHTS_TOGGLE.getGameBinding());
         }
     }